requirements.txt
Flask==3.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
openai==1.3.0
